requirements.txt
streamlit
requests
faiss-cpu
sentence-transformers
numpy
pandas
